+++ tests/models/test_hparams.py
@@ -41,14 +41,6 @@ class SaveHparamsModel(BoringModel):
         self.save_hyperparameters(hparams)
 
 
-class AssignHparamsModel(BoringModel):
-    """ Tests that a model can take an object with explicit setter """
-
-    def __init__(self, hparams):
-        super().__init__()
-        self.hparams = hparams
-
-
 def decorate(func):
 
     @functools.wraps(func)
@@ -66,16 +58,6 @@ class SaveHparamsDecoratedModel(BoringModel):
     def __init__(self, hparams, *my_args, **my_kwargs):
         super().__init__()
         self.save_hyperparameters(hparams)
-
-
-class AssignHparamsDecoratedModel(BoringModel):
-    """ Tests that a model can take an object with explicit setter"""
-
-    @decorate
-    @decorate
-    def __init__(self, hparams, *my_args, **my_kwargs):
-        super().__init__()
-        self.hparams = hparams
 
 
 # -------------------------
@@ -114,7 +96,7 @@ def _run_standard_hparams_test(tmpdir, model, cls, try_overwrite=False):
 
 
 @pytest.mark.parametrize(
-    "cls", [SaveHparamsModel, AssignHparamsModel, SaveHparamsDecoratedModel, AssignHparamsDecoratedModel]
+    "cls", [SaveHparamsModel, SaveHparamsDecoratedModel]
 )
 def test_namespace_hparams(tmpdir, cls):
     # init model
@@ -125,7 +107,7 @@ def test_namespace_hparams(tmpdir, cls):
 
 
 @pytest.mark.parametrize(
-    "cls", [SaveHparamsModel, AssignHparamsModel, SaveHparamsDecoratedModel, AssignHparamsDecoratedModel]
+    "cls", [SaveHparamsModel, SaveHparamsDecoratedModel]
 )
 def test_dict_hparams(tmpdir, cls):
     # init model
@@ -136,7 +118,7 @@ def test_dict_hparams(tmpdir, cls):
 
 
 @pytest.mark.parametrize(
-    "cls", [SaveHparamsModel, AssignHparamsModel, SaveHparamsDecoratedModel, AssignHparamsDecoratedModel]
+    "cls", [SaveHparamsModel, SaveHparamsDecoratedModel]
 )
 def test_omega_conf_hparams(tmpdir, cls):
     # init model
@@ -580,8 +562,7 @@ class SuperClassPositionalArgs(BoringModel):
 
     def __init__(self, hparams):
         super().__init__()
-        self._hparams = None  # pretend BoringModel did not call self.save_hyperparameters()
-        self.hparams = hparams
+        self._hparams = hparams  # pretend BoringModel did not call self.save_hyperparameters()
 
 
 class SubClassVarArgs(SuperClassPositionalArgs):
@@ -617,8 +598,6 @@ def test_init_arg_with_runtime_change(tmpdir, cls):
     assert model.hparams.running_arg == 123
     model.hparams.running_arg = -1
     assert model.hparams.running_arg == -1
-    model.hparams = Namespace(abc=42)
-    assert model.hparams.abc == 42
 
     trainer = Trainer(
         default_root_dir=tmpdir,
@@ -664,18 +643,11 @@ def test_model_save_hyper_parameters_interpolation_with_hydra(tmpdir):
 
         def __init__(self, args_0, args_1, args_2, kwarg_1=None):
             self.save_hyperparameters()
-            self.test_hparams()
-            config_file = f"{tmpdir}/hparams.yaml"
-            save_hparams_to_yaml(config_file, self.hparams)
-            self.hparams = load_hparams_from_yaml(config_file)
-            self.test_hparams()
-            super().__init__()
-
-        def test_hparams(self):
             assert self.hparams.args_0.log == "Something"
             assert self.hparams.args_1['cfg'].log == "Something"
             assert self.hparams.args_2[0].log == "Something"
             assert self.hparams.kwarg_1['cfg'][0].log == "Something"
+            super().__init__()
 
     with initialize(config_path="conf"):
         args_0 = compose(config_name="config")